requirements.txt
django
django-environ
drf-yasg
djangorestframework-simplejwt
django-filter
django-rest-passwordreset
sendgrid